Django/requirements.txt
pillow
django-crispy-forms